+++ adapters/mqtt-vertx-base/src/main/java/org/eclipse/hono/adapter/mqtt/MqttProtocolAdapterProperties.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2016, 2021 Contributors to the Eclipse Foundation
+ * Copyright (c) 2016, 2022 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -22,15 +22,10 @@ import org.eclipse.hono.config.ProtocolAdapterProperties;
 public class MqttProtocolAdapterProperties extends ProtocolAdapterProperties {
 
     /**
-     * The default number of milliseconds to wait for PUBACK.
-     */
-    protected static final int DEFAULT_COMMAND_ACK_TIMEOUT = 100;
-    /**
      * The amount of time (in milliseconds) to wait for a device to acknowledge receiving a command message.
      */
     protected static final long DEFAULT_SEND_MESSAGE_TO_DEVICE_TIMEOUT = 1000L; // ms
 
-    private int commandAckTimeout = DEFAULT_COMMAND_ACK_TIMEOUT;
     private long sendMessageToDeviceTimeout = DEFAULT_SEND_MESSAGE_TO_DEVICE_TIMEOUT;
 
     /**
@@ -47,46 +42,7 @@ public class MqttProtocolAdapterProperties extends ProtocolAdapterProperties {
      */
     public MqttProtocolAdapterProperties(final MqttProtocolAdapterOptions options) {
         super(options.adapterOptions());
-        setCommandAckTimeout(options.commandAckTimeout());
         setSendMessageToDeviceTimeout(options.sendMessageToDeviceTimeout());
-    }
-
-    /**
-     * Gets the waiting for acknowledgement timeout in milliseconds for commands published with QoS 1.
-     * <p>
-     * This timeout is used by the MQTT adapter for commands published with QoS 1. If there is no acknowledgement
-     * within this time limit, then the command is settled with the <em>released</em> outcome.
-     * <p>
-     * The default value is {@link #DEFAULT_COMMAND_ACK_TIMEOUT}.
-     *
-     * @deprecated Use {@link #getSendMessageToDeviceTimeout()} instead.
-     *
-     * @return The timeout in milliseconds.
-     */
-    @Deprecated(forRemoval = true)
-    public final int getCommandAckTimeout() {
-        return commandAckTimeout;
-    }
-
-    /**
-     * Sets the waiting for acknowledgement timeout in milliseconds for commands published with QoS 1.
-     * <p>
-     * This timeout is used by the MQTT adapter for commands published with QoS 1. If there is no acknowledgement
-     * within this time limit, then the command is settled with the <em>released</em> outcome.
-     * <p>
-     * The default value is {@link #DEFAULT_COMMAND_ACK_TIMEOUT}.
-     *
-     * @deprecated Use {@link #setSendMessageToDeviceTimeout(long)} ()} instead.
-     *
-     * @param timeout The timeout in milliseconds.
-     * @throws IllegalArgumentException if the timeout is negative.
-     */
-    @Deprecated(forRemoval = true)
-    public final void setCommandAckTimeout(final int timeout) {
-        if (timeout < 0) {
-            throw new IllegalArgumentException("timeout must not be negative");
-        }
-        this.commandAckTimeout = timeout;
     }
 
     /**
@@ -121,22 +77,4 @@ public class MqttProtocolAdapterProperties extends ProtocolAdapterProperties {
 
         this.sendMessageToDeviceTimeout = sendMessageToDeviceTimeout;
     }
-
-    /**
-     * Gets the effective timeout for waiting for acknowledgement in milliseconds for commands published with QoS 1
-     * by taking the {@link #sendMessageToDeviceTimeout} and {@link #commandAckTimeout} properties into account.
-     *
-     * Can be removed when the deprecated {@link #commandAckTimeout} property is removed.
-     *
-     * @return The timeout in milliseconds.
-     */
-    long getEffectiveSendMessageToDeviceTimeout() {
-        if (sendMessageToDeviceTimeout == DEFAULT_SEND_MESSAGE_TO_DEVICE_TIMEOUT
-                && commandAckTimeout != DEFAULT_COMMAND_ACK_TIMEOUT) {
-            return commandAckTimeout;
-        } else {
-            return sendMessageToDeviceTimeout;
-        }
-    }
-
 }